Add files via upload
In a Heterogenous Feature Selection Ensemble, multiple AI feature selectors  are chosen to be a part of this ensemble as the name implies. These AI feature selectors in this ensemble vote for the best features using a ranking score. The feature that receives the highest votes is considered the best while the feature with the lowest votes is the least significant. Using a variety of AI models for selecting features is advantageous because it controls the variance, and reduces the likelihood of poor feature selection.